Added initial setup for 3.0.9

+++ includes/class-epofw-field-table.php
@@ -8,6 +8,7 @@
  * @package    Extra_Product_Options_For_WooCommerce
  * @subpackage Extra_Product_Options_For_WooCommerce/includes
  */
+
 // If this file is called directly, abort.
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -36,7 +37,7 @@ if ( ! class_exists( 'EPOFW_Field_Table' ) ) {
 		/**
 		 * Get post type
 		 *
-		 * @var $post_type store post type.
+		 * @var $post_type string store post type.
 		 *
 		 * @since 1.0.0
 		 */
@@ -120,33 +121,36 @@ if ( ! class_exists( 'EPOFW_Field_Table' ) ) {
 		public function prepare_items() {
 			$this->prepare_column_headers();
 			$per_page    = $this->get_items_per_page( 'epofw_per_page' );
-			$get_search  = filter_input( INPUT_POST, 's', FILTER_SANITIZE_STRING );
-			$get_orderby = filter_input( INPUT_GET, 'orderby', FILTER_SANITIZE_STRING );
-			$get_order   = filter_input( INPUT_GET, 'order', FILTER_SANITIZE_STRING );
+			$get_search  = filter_input( INPUT_POST, 's', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+			$get_search  = isset( $get_search ) ? sanitize_text_field( wp_unslash( $get_search ) ) : '';
+			$get_orderby = filter_input( INPUT_GET, 'orderby', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+			$get_orderby = isset( $get_orderby ) ? sanitize_text_field( wp_unslash( $get_orderby ) ) : '';
+			$get_order   = filter_input( INPUT_GET, 'order', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+			$get_order   = isset( $get_order ) ? sanitize_text_field( wp_unslash( $get_order ) ) : '';
 			$args        = array(
 				'posts_per_page' => $per_page,
 				'orderby'        => 'ID',
 				'order'          => 'DESC',
 				'offset'         => ( $this->get_pagenum() - 1 ) * $per_page,
 			);
-			if ( isset( $get_search ) && ! empty( $get_search ) ) {
+			if ( ! empty( $get_search ) ) {
 				$args['s'] = trim( wp_unslash( $get_search ) );
 			}
-			if ( isset( $get_orderby ) && ! empty( $get_orderby ) ) {
+			if ( ! empty( $get_orderby ) ) {
 				if ( 'title' === $get_orderby ) {
 					$args['orderby'] = 'title';
 				} elseif ( 'date' === $get_orderby ) {
 					$args['orderby'] = 'date';
 				}
 			}
-			if ( isset( $get_order ) && ! empty( $get_order ) ) {
+			if ( ! empty( $get_order ) ) {
 				if ( 'asc' === strtolower( $get_order ) ) {
 					$args['order'] = 'ASC';
 				} elseif ( 'desc' === strtolower( $get_order ) ) {
 					$args['order'] = 'DESC';
 				}
 			}
-			$this->items = $this->epofw_find( $args, $get_orderby );
+			$this->items = $this->epofw_find( $get_orderby, $args );
 			$total_items = $this->epofw_count();
 			$total_pages = ceil( $total_items / $per_page );
 			$this->set_pagination_args(
@@ -163,20 +167,17 @@ if ( ! class_exists( 'EPOFW_Field_Table' ) ) {
 		 * @since 1.0.0
 		 */
 		public function no_items() {
-			if ( isset( $this->error ) ) {
-				echo esc_html( $this->error->get_error_message() );
-			} else {
-				esc_html_e( 'No Field Option found.', 'extra-product-options-for-woocommerce' );
-			}
+			esc_html_e( 'No Field Option found.', 'extra-product-options-for-woocommerce' );
 		}
+
 		/**
 		 * Checkbox column.
 		 *
-		 * @param string $item Get Field Option id.
-		 *
-		 * @return mixed
-		 *
 		 * @since 1.0.0
+		 *
+		 * @param object $item Get Field Option id.
+		 *
+		 * @return string|void
 		 */
 		public function column_cb( $item ) {
 			if ( ! $item->ID ) {
@@ -228,9 +229,9 @@ if ( ! class_exists( 'EPOFW_Field_Table' ) ) {
 			$delurl               = self::$admin_object->dynamic_url( self::$current_page, self::$current_tab, 'delete', $item->ID );
 			$duplicateurl         = self::$admin_object->dynamic_url( self::$current_page, self::$current_tab, 'duplicate', $item->ID );
 			$actions              = array();
-			$actions['edit']      = '<a href="' . wp_nonce_url( $editurl, 'edit_' . esc_attr( $item->ID ), 'epofw_nonce' ) . '">' . esc_html__( 'Edit', 'extra-product-options-for-woocommerce' ) . '</a>';
-			$actions['delete']    = '<a href="' . wp_nonce_url( $delurl, 'del_' . esc_attr( $item->ID ), 'epofw_nonce' ) . '">' . esc_html__( 'Delete', 'extra-product-options-for-woocommerce' ) . '</a>';
-			$actions['duplicate'] = '<a href="' . wp_nonce_url( $duplicateurl, 'duplicate_' . esc_attr( $item->ID ), 'epofw_nonce' ) . '">' . esc_html__( 'Duplicate', 'extra-product-options-for-woocommerce' ) . '</a>';
+			$actions['edit']      = '<a href="' . wp_nonce_url( $editurl, 'edit_' . $item->ID, 'epofw_nonce' ) . '">' . esc_html__( 'Edit', 'extra-product-options-for-woocommerce' ) . '</a>';
+			$actions['delete']    = '<a href="' . wp_nonce_url( $delurl, 'del_' . $item->ID, 'epofw_nonce' ) . '">' . esc_html__( 'Delete', 'extra-product-options-for-woocommerce' ) . '</a>';
+			$actions['duplicate'] = '<a href="' . wp_nonce_url( $duplicateurl, 'duplicate_' . $item->ID, 'epofw_nonce' ) . '">' . esc_html__( 'Duplicate', 'extra-product-options-for-woocommerce' ) . '</a>';
 			return $this->row_actions( $actions );
 		}
 		/**
@@ -289,16 +290,23 @@ if ( ! class_exists( 'EPOFW_Field_Table' ) ) {
 		 * @since 1.0.0
 		 */
 		public function process_bulk_action() {
-			$delete_nonce     = filter_input( INPUT_POST, '_wpnonce', FILTER_SANITIZE_STRING );
-			$get_method_id_cb = filter_input( INPUT_POST, 'method_id_cb', FILTER_SANITIZE_NUMBER_INT, FILTER_REQUIRE_ARRAY );
-			$method_id_cb     = ! empty( $get_method_id_cb ) ? array_map( 'sanitize_text_field', wp_unslash( $get_method_id_cb ) ) : array();
-			$get_tab          = filter_input( INPUT_GET, 'tab', FILTER_SANITIZE_STRING );
-			$action           = $this->current_action();
-			if ( ! isset( $method_id_cb ) ) {
+			$delete_nonce = filter_input( INPUT_POST, '_wpnonce', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+			$delete_nonce = isset( $delete_nonce ) ? sanitize_text_field( wp_unslash( $delete_nonce ) ) : '';
+			if ( empty( $delete_nonce ) || ! wp_verify_nonce( $delete_nonce, 'bulk-shippingmethods' ) ) {
 				return;
 			}
-			$deletenonce = wp_verify_nonce( $delete_nonce, 'bulk-shippingmethods' );
-			if ( ! isset( $deletenonce ) && 1 !== $deletenonce ) {
+			$get_method_id_cb = filter_input( INPUT_POST, 'method_id_cb', FILTER_SANITIZE_NUMBER_INT, FILTER_REQUIRE_ARRAY );
+			$method_id_cb     = ! empty( $get_method_id_cb ) ? array_map( 'sanitize_text_field', wp_unslash( $get_method_id_cb ) ) : array();
+			if ( empty( $method_id_cb ) ) {
+				return;
+			}
+			$get_tab = filter_input( INPUT_GET, 'tab', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+			$get_tab = isset( $get_tab ) ? sanitize_text_field( wp_unslash( $get_tab ) ) : '';
+			if ( empty( $get_tab ) ) {
+				$get_tab = 'general';
+			}
+			$action = $this->current_action();
+			if ( ! in_array( $action, array( 'delete', 'enable', 'disable' ), true ) ) {
 				return;
 			}
 			$items = array_filter( array_map( 'absint', $method_id_cb ) );
@@ -306,28 +314,34 @@ if ( ! class_exists( 'EPOFW_Field_Table' ) ) {
 				return;
 			}
 			if ( 'delete' === $action ) {
-				foreach ( $items as $id ) {
-					wp_delete_post( $id );
+				if ( ! empty( $items ) ) {
+					foreach ( $items as $id ) {
+						wp_delete_post( $id );
+					}
 				}
 				self::$admin_object->epofw_updated_message( 'deleted', $get_tab, '' );
 			} elseif ( 'enable' === $action ) {
-				foreach ( $items as $id ) {
-					$enable_post = array(
-						'post_type'   => self::$post_type,
-						'ID'          => $id,
-						'post_status' => 'publish',
-					);
-					wp_update_post( $enable_post );
+				if ( ! empty( $items ) ) {
+					foreach ( $items as $id ) {
+						$enable_post = array(
+							'post_type'   => self::$post_type,
+							'ID'          => $id,
+							'post_status' => 'publish',
+						);
+						wp_update_post( $enable_post );
+					}
 				}
 				self::$admin_object->epofw_updated_message( 'enabled', $get_tab, '' );
 			} elseif ( 'disable' === $action ) {
-				foreach ( $items as $id ) {
-					$disable_post = array(
-						'post_type'   => self::$post_type,
-						'ID'          => $id,
-						'post_status' => 'draft',
-					);
-					wp_update_post( $disable_post );
+				if ( ! empty( $items ) ) {
+					foreach ( $items as $id ) {
+						$disable_post = array(
+							'post_type'   => self::$post_type,
+							'ID'          => $id,
+							'post_status' => 'draft',
+						);
+						wp_update_post( $disable_post );
+					}
 				}
 				self::$admin_object->epofw_updated_message( 'disabled', $get_tab, '' );
 			}
@@ -335,15 +349,14 @@ if ( ! class_exists( 'EPOFW_Field_Table' ) ) {
 		/**
 		 * Find post data.
 		 *
-		 * @param mixed  $args        pass query args.
-		 *
 		 * @param string $get_orderby pass order by for listing.
+		 * @param mixed  $args        pass query args.
 		 *
 		 * @return array $posts
 		 *
 		 * @since 1.0.0
 		 */
-		public static function epofw_find( $args = '', $get_orderby ) {
+		public static function epofw_find( $get_orderby, $args = '' ) {
 			$defaults          = array(
 				'post_status'    => 'any',
 				'posts_per_page' => - 1,
@@ -355,23 +368,27 @@ if ( ! class_exists( 'EPOFW_Field_Table' ) ) {
 			$args['post_type'] = self::$post_type;
 			$epofw_query       = new WP_Query( $args );
 			$posts             = $epofw_query->query( $args );
-			if ( ! isset( $get_orderby ) && empty( $get_orderby ) ) {
+			if ( ! isset( $get_orderby ) ) {
 				$sort_order     = array();
 				$get_sort_order = get_option( 'sm_sortable_order' );
-				if ( isset( $get_sort_order ) && ! empty( $get_sort_order ) ) {
+				if ( ! empty( $get_sort_order ) ) {
 					foreach ( $get_sort_order as $sort ) {
 						$sort_order[ $sort ] = array();
 					}
 				}
-				foreach ( $posts as $carrier_id => $carrier ) {
-					$carrier_name = $carrier->ID;
-					if ( array_key_exists( $carrier_name, $sort_order ) ) {
-						$sort_order[ $carrier_name ][ $carrier_id ] = $posts[ $carrier_id ];
-						unset( $posts[ $carrier_id ] );
+				if ( ! empty( $posts ) ) {
+					foreach ( $posts as $carrier_id => $carrier ) {
+						$carrier_name = $carrier->ID;
+						if ( array_key_exists( $carrier_name, $sort_order ) ) {
+							$sort_order[ $carrier_name ][ $carrier_id ] = $posts[ $carrier_id ];
+							unset( $posts[ $carrier_id ] );
+						}
 					}
 				}
-				foreach ( $sort_order as $carriers ) {
-					$posts = array_merge( $posts, $carriers );
+				if ( ! empty( $sort_order ) ) {
+					foreach ( $sort_order as $carriers ) {
+						$posts = array_merge( $posts, $carriers );
+					}
 				}
 			}
 			self::$epofw_found_items = $epofw_query->found_posts;
